[New] Application of symmetry for quadratic boson Hamiltonian
- The Hamiltonian used in, e.g., spinW consists of four N by N
  block matrices combined into a single 2N by 2N matrix, for a magnetic
  system with N symmetry distinct ions. As a result the eigenvectors of
  the Hamiltonian matrix contain 2N elements partitioned into two 0:N
  sub-vectors indexed by the ion indexes.
  If the system contains more than one magnetic ion and has a symmetry
  above P1 then there will be an ion-index permutation for each symmetry
  operation. After interpolating in the irreducible Brillouin zone, the
  inverse symmetry operation's ion permutation must be used to permute
  the interpolated eigenvector elements.
- The eigenvector element permutation has been implemented by specifying
  a new `RotatesLike` enumerated value `Copla` to indicate the new
  behaviour. Setting this value (4 from the Python grid `fill` methods)
  causes the `PhononGamma` table to be calculated when
  `ir_interpolate_at` is called [And every time it's called].
  The `PhononGamma` table contains site-equivalency permutation
  information for every symmetry operation consisting of equal-site
  index mapping and an integer lattice vector if the operation moves a
  site from one unit cell to another. The new in-place
  post-interpolation rotation method for `RotatesLike::Copla` uses the
  index mapping, so there may be performance benefits in splitting
  `PhononGamma` into a superclass containing the index map and a
  subclass adding the vector map, for expediency this has not been
  investigated.
- This implementation requires a careful correspondence of `Direct`
  lattice basis and eigenvectors, namely that the basis can only contain
  the *magnetic* ions and the eigenvectors must have twice as many
  entries as magnetic ions.
  There may also be pitfalls in ensuring that the magnetic lattice is
  used as it will have a different symmetry from the crystallographic
  lattice in some cases.

[Fix] compiler warnings

+++ src/permutation.hpp
@@ -352,7 +352,8 @@ brille::ind_t nind = static_cast<brille::ind_t>(neighbour_idx);
 brille::ind_t cind = static_cast<brille::ind_t>(centre_idx);
 for (brille::ind_t i=0; i<Nobj; ++i)
   for (brille::ind_t j=0; j<Nobj; ++j)
-    if (permutations.val(nind, i) == rowsol[j]) permutations.val(cind, i) = static_cast<size_t>(j);
+    if (permutations.val(nind, i) == static_cast<size_t>(rowsol[j]))
+      permutations.val(cind, i) = static_cast<size_t>(j);
 
 delete[] cost;
 delete[] usol;
@@ -541,7 +542,8 @@ brille::ind_t nind = static_cast<brille::ind_t>(neighbour_idx);
 brille::ind_t cind = static_cast<brille::ind_t>(centre_idx);
 for (brille::ind_t i=0; i<Nobj; ++i)
   for (brille::ind_t j=0; j<Nobj; ++j)
-    if (permutations.val(nind, i) == rowsol[j]) permutations.val(cind, i) = static_cast<size_t>(j);
+    if (permutations.val(nind, i) == static_cast<size_t>(rowsol[j]))
+      permutations.val(cind, i) = static_cast<size_t>(j);
 
 delete[] cost;
 delete[] usol;
